requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
boto3==1.40.41
botocore==1.40.41
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
circlify==0.15.1
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
httplib2==0.31.0
idna==3.10
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
narwhals==2.10.2
networkx==3.5
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.1
psycopg2==2.9.11
py4j==0.10.9.7
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
pyspark==3.5.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.28.0
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
smmap==5.0.2
streamlit==1.51.0
streamlit-plotly-events==0.0.6
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wordcloud==1.9.4
